fix type of weight

+++ src/main/java/model/Vegetable.java
@@ -3,10 +3,10 @@ package model;
 public class Vegetable extends FoodIngredient{
     private VegetableType type;
 
-    public Vegetable(VegetableType type,int weight){
+        public Vegetable(VegetableType type,double weight){
         this.type = type;
         this.weight = weight;
-        calories = type.caloriesPerKg * weight;
+        calories = (int)(type.caloriesPerKg * weight);
     }
 
     public VegetableType getType(){
